Set the PSA key type when parsing a private key
In mbedtls_pk_parse_key(), set the PSA key type based on the internal
PSA private key, once we have finished parsing.

For simplicity, do this just before we return on the success path.
However, since there are a lot of points in mbedtls_pk_parse_key() where
we return on the success path, make a static helper function and call
that at all of those points.

Signed-off-by: David Horstmann <david.horstmann@arm.com>