Set the PSA key type for public key PK objects
When parsing a public key, set the PSA key type to be the correct type
for the public key. This is done separately for RSA and ECC keys.

Note that this is the PSA type that the public key would have if it were
stored in PSA (however, it is not).

Signed-off-by: David Horstmann <david.horstmann@arm.com>