drivers|tests: rsa: check bit-size on key pair import
This fix is required to make the CI happy on test
'PSA import RSA key pair: maximum size exceeded' from
'test_suite_psa_crypto.data'. As explained in the comments added in this
commit, issue tf-psa-crypto#562 has been created to track this testing
weakness and to solve it in the future.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>