Change RSA key sizes to 1024
This allows the test cases to run when we are testing the static key
slots in the test_crypto_with_static_key_slots component. That component
deliberately sets a static key slot size that prevents key pairs larger
than RSA 1024.

We could try and gate the tests based on whether we are testing static
key slots but it seems better to retain the coverage. The number of bits
is a parameter of the testcase, so it's easy to change if we change our
mind.

Signed-off-by: David Horstmann <david.horstmann@arm.com>