Fix incorrect calculation of ASN1 max size
For bit lengths that are not a multiple of 8, the macro
PSA_KEY_EXPORT_ASN1_INTEGER_MAX_SIZE calculated the incorrect result (it
did integer division by 8 rather than an integer ceil() operation).

This would be fine if we were always using multiples of 8, however to
calculate the maximum size of RSA keys we use a non-multiple of 8 to
allow extra room. This should generate an extra byte in the size.

Signed-off-by: David Horstmann <david.horstmann@arm.com>