Move crypto_adjust_config_auto_enabled.h where it belongs
This commit changes the location of the file in the inclusion order, to
group it with other headers that tweak the PSA configuration. The former
order also worked because the macros that it references aren't used in
other configuration-tweaking headers at this time (and have never been,
I believe), so this is not a bug fix, just a readability and robustness
improvement.

This is at the PSA level, not driver-specific, so keep it in `/include`.

This is not meant to be included directly by applications, so move it under
`.../private`.

Rename the file to match the pattern followed by similar files (I had
forgotten a word when I created that file initially).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>