Move crypto_adjust_config_derived.h where it belongs
This commit changes the location of the file in the inclusion order, to
group it with other headers that tweak the PSA configuration. Specificaly,
this header is now included before `config_adjust_legacy_from_psa.h` and
`config_adjust_test_accelerators.h`. Both of these set driver-specific
macros based on `PSA_WANT` macros, whereas
`crypto_adjust_config_derived.h` does not query driver-specific macros, so
the new order is no worse.

In fact the new order more correct because it would make sense for
`config_adjust_legacy_from_psa.h` to query macros set by
`crypto_adjust_config_derived.h`. However, it currently doesn't, so this
is not a bug fix, just a robustness improvement.

This is at the PSA level, not driver-specific, so keep it in `/include`.

This is not meant to be included directly by applications, so move it under
`.../private`.

Rename the file to match the pattern followed by similar files (I had
forgotten a word when I created that file initially).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>