Move config_adjust_legacy_crypto.h where it belongs
This is specific to modules that are under `/drivers`, so keep it under
`/drivers`. Some of these modules are also public interfaces that are used
in builds with only third-party drivers, so the corresponding macros should
be outside of `/drivers`; this will be handled in a subsequent commit.

This is not meant to be included directly by applications, so move it under
`.../private`.

Rename the file to match the pattern followed by similar files, and to
better reflect its current role: it's tweaking modules that implement the
built-in cryptography, which were formerly legacy public interfaces but are
now internal.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>