Switch approach for setting psa_type
For key pairs, switch approach for setting the PSA key type. Instead of
setting the key type in many arbitrary places, set it in only two places
(when parsing key pairs):
* mbedtls_pk_rsa_set_key(), for RSA key pairs
* mbedtls_pk_ecc_set_key(), for EC key pairs

This limits the complexity and reduces the number of places where the
key type is assigned, so that it's easier to ensure we do not miss one.

Signed-off-by: David Horstmann <david.horstmann@arm.com>