Add a parameter to mbedtls_pk_ecc_set_pubkey()
This function is sometimes used to build an ordinary public key and
sometimes used to fill in the public part of a key pair, so it cannot
reliably set the PSA key type.

As a (perhaps clumsy) fix to this, add a parameter to specify whether
the function is being used to fill in the public part of an existing key
pair (is_keypair) and use that to decide whether to update the PSA key
type.

Signed-off-by: David Horstmann <david.horstmann@arm.com>