Remove enforcement against direct inclusion of private headers
We added `#error` directives to prevent some headers from being included
directly because those headers were not clearly indicated as public (and, in
the case of `check_config.h`, were formerly public). Headers under
`.../private` have no history of being public and are obviously not intended
for direct consumption, so remove the enforcement.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>