Remove is_keypair parameter
Instead of explicitly specifying whether we are setting the public part
of a keypair, look to see whether the PSA type has been set already. If
it has been set, leave the existing key type in place as long as it
is the keypair type corresponding to the public key we are setting.
Otherwise, return an error.

Co-authored-by: Valerio Setti <vsetti@baylibre.com>
Signed-off-by: David Horstmann <david.horstmann@arm.com>