drivers: pk: enhance mbedtls_pk_write_pubkey_psa
Allow the function to be able to return the public key in all the scenarios
supported by PK:
- If the public key is already available in the PK context, then simply
  memcpy() it.
- If only the private key is available, then export the public key from
  the private one and then memcpy() it.

Advantages are:
- The user doesn't need to worry about filling the public part before
  calling this function.
- It allows the function to work also with opaque keys. This functionality
  wasn't supported before, so it's an improvement that comes for free.

The only side effect of this change is that the first parameter of
mbedtls_pk_write_pubkey_psa (i.e. mbedtls_pk_context) cannot be const
because otherwise it wouldn't be possible to pass it to
"mbedtls_pk_set_pubkey_from_prv" without some dirty casting.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>