drivers: pk: pk_parse_key: RSA: export public key when parsing private key
For the EC keys this is already happening in "pk_parse_key_sec1_der()"
even though the procedure is a bit different in that case. For SEC1
the key might or might not be present on the parsed buffer, so it might
or might not be derived form the private counterpart. For RSA keys there
is no such possibility: the public key should always be exported from the
private key.

The point here is that at the end of the "mbedtls_pk_parse_key()" we want
private and public fields of the PK context to be filled for all the
supported key types.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>