drivers: pkwrite: do not modify PK context in mbedtls_pk_write_pubkey_psa
Partially reverts commit 63660b73a4cd0c3e02a0edc3c7bacb913dfaee34 for
what concerns with the following:

- keep "const" on the PK parameter being passed in input to the function;
- do not try to complete the PK context if the public key is missing.

The rationale here is that this is just a debug function meant to get some
data from the PK context. It's the caller responsibility to properly fill
the PK context before calling it.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>